// server/src/types/register.typer.ts
import { t } from "elysia";

export const _register = t.Object({
    username: t.String(),
    password: t.String(),
    display_name: t.String(),
    dete_of_birth: t.Optional(t.Date()),
    looking_for: t.Union ([t.Literal ('male'),t.Literal('famal'),t.Literal('all')]),
    gender: t.Optional(t.Union([t.Literal ('male'),t.Literal('famal'),t.Literal('all')]))
})